Revert efforts for running gevent tests on Travis
We could possibly install libevent-2.0-5 using apt-get, and then gevent and
greenlet using pip, and make the tests pass for Python 2.x. Though, I can't
find any way to block the gevent and greenlet installation to be executed for
Python 3 as well, and it would thus break the Python 3 build.