Let actors access themselves through a proxy
Fixes #9

An actor can use this to schedule work for itself. The scheduled work will only
be done after the current message and all messages already in the inbox are
processed.

For example, if an actor can split a time consuming task into multiple parts,
and after completing each part can ask itself to start on the next part using
proxied calls or messages to itself, it can react faster to other incoming
messages as they will be interleaved with the parts of the time consuming task.
This is especially useful for being able to stop the actor in the middle of a
time consuming task.